feat: enhance tab bar styling with transparent background and custom create button
- Add transparent background for iOS tab bar with blur effect
- Implement custom create button with AntDesign plus icon and colored circle
- Make active and inactive tab colors consistent across light/dark modes
- Adjust create button positioning and sizing for better visual hierarchy
- Reduce blur intensity for more subtle transparency effect
