+++ components/ui/TabBarBackground.ios.tsx
@@ -8,7 +8,7 @@ export default function BlurTabBarBackground() {
       // System chrome material automatically adapts to the system's theme
       // and matches the native tab bar appearance on iOS.
       tint="systemChromeMaterial"
-      intensity={100}
+      intensity={75}
       style={StyleSheet.absoluteFill}
     />
   );
