+++ constants/Colors.ts
@@ -14,6 +14,8 @@ export const Colors = {
     icon: "#687076",
     tabIconDefault: "#687076",
     tabIconSelected: tintColorLight,
+    tabBarButtonColor: "#0C0C0C",
+    tabBarButtonColorBg: "#EDEDED",
     appPrimary: "#F01E21",
     success: "#4CAF50",
     error: "#F44336",
@@ -23,6 +25,8 @@ export const Colors = {
     background: "#0F0F0F",
     tint: tintColorDark,
     icon: "#9BA1A6",
+    tabBarButtonColor: "#FFFFFF",
+    tabBarButtonColorBg: "#2E2E2E",
     tabIconDefault: "#9BA1A6",
     tabIconSelected: tintColorDark,
     appPrimary: "#F01E21",
